fix: Only pull entire list of artifacts once
If you have multiple "rules" it would previously cause the code to pull
the entire list of artifacts in all the repos, every time.

This optimization will keep that list and only pull it once per repo.

In our environment with ~1.5M artifacts in the repo, this saves ~3mins per rule.

~tommy
Co-authored with Nithya (nittyy)